# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Generar Token de Acceso (Login)",
                "description": "Autentica al usuario con correo, contraseña y dispositivo para obtener un token de acceso.",
                "operationId": "1a8d4a3aa356c539118bca7ed8b657a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "correo",
                                    "contraseña",
                                    "dispositivo"
                                ],
                                "properties": {
                                    "correo": {
                                        "description": "Correo del usuario.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "Erwin@example.com"
                                    },
                                    "contraseña": {
                                        "description": "Contraseña del usuario.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "dispositivo": {
                                        "description": "Nombre del dispositivo (e.g., 'windows', 'ios', 'android').",
                                        "type": "string",
                                        "example": "windows"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso. Retorna un token de acceso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "attributes": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "nombre": {
                                                            "type": "string",
                                                            "example": "Erwin Santiago Arrega Avila"
                                                        },
                                                        "correo": {
                                                            "type": "string",
                                                            "format": "email",
                                                            "example": "Erwin@example.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "description": "Token de acceso JWT.",
                                                    "type": "string",
                                                    "example": "lkYTYtksh64dJh9bqKrFa5GHpQZax2IA9D8fdI7Rc54a30ac"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales no válidas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciales no válidas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/membresias": {
            "get": {
                "tags": [
                    "Membresias"
                ],
                "summary": "Consultar todas las membresias",
                "description": "Retorna todas las membresias",
                "operationId": "1cd9a7dcaab60995bc8ca0bf136a68ab",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No se encontraron membresias"
                    },
                    "405": {
                        "description": "Método no permitido"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Membresias"
                ],
                "summary": "Crear membresias",
                "description": "Crear una nueva membresia",
                "operationId": "344c0c800105033ea8085d518aaed42f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "tipo",
                                    "costo",
                                    "detalles",
                                    "duracion_dias"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Membresia Oro"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Mensual"
                                    },
                                    "costo": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "49.99"
                                    },
                                    "detalles": {
                                        "type": "string",
                                        "example": "Acceso completo a todas las instalaciones y clases."
                                    },
                                    "duracion_dias": {
                                        "type": "integer",
                                        "example": "30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Membresia creada"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/membresias/{id}": {
            "get": {
                "tags": [
                    "Membresias"
                ],
                "summary": "Consultar una membresia por ID",
                "description": "Retorna una membresia específica",
                "operationId": "c1f76712b5c4de4fb013b6c4694cae35",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la membresía",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Membresia no encontrada"
                    },
                    "405": {
                        "description": "Método no permitido"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Membresias"
                ],
                "summary": "Actualizar membresias",
                "description": "Actualizar una membresia existente",
                "operationId": "5b09986fbdbbcf129d27798f6eb378e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la membresía a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "nombre",
                                    "tipo",
                                    "costo",
                                    "detalles",
                                    "duracion_dias"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "Debe ser 'PUT' o 'PATCH'",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Membresia Platino"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Anual"
                                    },
                                    "costo": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "499.99"
                                    },
                                    "detalles": {
                                        "type": "string",
                                        "example": "Acceso premium con entrenador personal y sauna."
                                    },
                                    "duracion_dias": {
                                        "type": "integer",
                                        "example": "365"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Membresia actualizada"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Membresia no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Membresias"
                ],
                "summary": "Eliminar membresia",
                "description": "Elimina una membresía existente por su ID.",
                "operationId": "433d66fa69b76f24e0b39040b6a624c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la membresía a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Membresía eliminada correctamente (Sin Contenido)"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Membresía no encontrada"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}